Added Expand Collapse working
